Feat: 요일,날짜를 계산하는 코드를 커스텀훅으로 분리 #47
